requirements.txt
pandas
sqlalchemy
psycopg2-binary
numpy
scipy
scikit-learn
matplotlib
seaborn
jupyter
python-dotenv
openpyxl
